requirements.txt
gradio>=4.0.0
pandas>=2.0.0
llama-cpp-python>=0.2.0
requests>=2.31.0
pydantic>=2.0.0